requirements.txt
Django>=4.2
djangorestframework>=3.15
mysqlclient>=2.1
python-dotenv>=1.0
djangorestframework-simplejwt>=5.2.0
django-filter>=23.0
